Refactored visualizations data filter code
